fix(gpu/cuda,build): get Windows backend compiling, bump 0.1.12
- include_shader! cuda: include_str! -> include_bytes! to match the &[u8]
  declared by declare_kernel! (was &str, mismatched the const type).
- cuda/buffer::copy_buffer: drop bogus `as usize` casts; CUdeviceptr is u64
  and src/dst_offset are already u64, wrapping_add expects u64.
- cuda/pipeline::load_kernel: tighten fname to &'static str so the cache key
  (HashMap<(usize, &'static str), KernelEntry>) cannot escape; documented
  the constraint on the public function and in declare_kernel!'s docstring.
- cuda/pipeline::load_module_and_func: strip slangc's trailing NUL on
  Windows before CString::new (was a NulError in kernel code on first
  dispatch).
- build/compile: wrap each slang CPU TU body in an anonymous namespace so
  vekl helpers (FromRGBA_0, LoadPixel_0, Gaussian1d_0, ...) become TU-local;
  fixes LNK2005 in any crate with >=2 shaders that share vekl (vignette).
  Includes leading-#line guard and 4 unit tests.

+++ src/gpu/backends/cuda/buffer.rs
@@ -191,8 +191,8 @@ pub unsafe fn copy_buffer(
 	let ctx = device as cudarc::driver::sys::CUcontext;
 	unsafe { cuCtxSetCurrent(ctx) };
 
-	let src_dev = (src as CUdeviceptr).wrapping_add(src_offset as usize);
-	let dst_dev = (dst as CUdeviceptr).wrapping_add(dst_offset as usize);
+	let src_dev = (src as CUdeviceptr).wrapping_add(src_offset);
+	let dst_dev = (dst as CUdeviceptr).wrapping_add(dst_offset);
 
 	if src_pitch_bytes == dst_pitch_bytes && src_pitch_bytes == width_bytes {
 		let total = (width_bytes as usize).saturating_mul(height as usize);
